fix: rename home sections
