+++ src/components/Home/Home.tsx
@@ -1,8 +1,7 @@
-import React from "react";
+import React, { useEffect } from "react";
 
 import { home } from "../../content/index";
 import { Categories, Footer, HomeHeader, Partners, Projects } from "./index";
-import { useEffect } from "react";
 
 export const Home = () => {
   const headerData = home.header;
