+++ src/components/Home/HomeSections/Categories/Categories.styles.scss
