+++ src/components/Home/HomeSections/Partners/Partner.tsx
