+++ src/components/Home/HomeSections/Partners/Partners.styles.scss
