+++ src/components/Home/HomeSections/Partners/Partners.tsx
