+++ src/components/Home/HomeSections/Projects/Carrousel/Carrousel.styles.scss
