+++ src/components/Home/HomeSections/Projects/Carrousel/CarrouselSlide.tsx
