+++ src/components/Home/HomeSections/Projects/Projects.styles.scss
