+++ src/components/Home/index.ts
@@ -1,7 +1,7 @@
 export * from "./HomeHeader/HomeHeader";
-export * from "./HomeSections/1-categories/Categories";
-export * from "./HomeSections/2-projects/Projects";
-export * from "./HomeSections/3-partners/Partners";
+export * from "./HomeSections/Categories/Categories";
+export * from "./HomeSections/Projects/Projects";
+export * from "./HomeSections/Partners/Partners";
 export * from "./AllowCookies/AllowCookies";
 
 export * from "../shared/Nav/Nav";
